# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrendDirection =
  (typeof TrendDirection)[keyof typeof TrendDirection];

export const TrendDirection = {
  up: "up",
  down: "down",
  flat: "flat",
} as const;

export type StatCardFormat =
  (typeof StatCardFormat)[keyof typeof StatCardFormat];

export const StatCardFormat = {
  currency: "currency",
  number: "number",
  percent: "percent",
} as const;

export interface StatCard {
  label: string;
  value: number;
  format: StatCardFormat;
  deltaPercent: number;
  trend: TrendDirection;
  positiveIsGood: boolean;
}

export interface OverviewMetrics {
  mrr: StatCard;
  activeUsers: StatCard;
  churnRate: StatCard;
  avgRevenue: StatCard;
}

export interface RevenuePoint {
  /** YYYY-MM */
  month: string;
  /** Short month label e.g. Jan */
  label: string;
  revenue: number;
}

export interface PlanBreakdownItem {
  plan: string;
  customers: number;
  revenue: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  signup: "signup",
  payment: "payment",
  churn: "churn",
  upgrade: "upgrade",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  customerName: string;
  amount?: number | null;
  occurredAt: string;
}

export type CustomerStatus =
  (typeof CustomerStatus)[keyof typeof CustomerStatus];

export const CustomerStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Customer {
  id: number;
  name: string;
  email: string;
  status: CustomerStatus;
  plan: string;
  mrr: number;
  createdAt: string;
}

export interface CreateCustomerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  status: CustomerStatus;
  plan: string;
  mrr: number;
}

export type RevenueLogType =
  (typeof RevenueLogType)[keyof typeof RevenueLogType];

export const RevenueLogType = {
  subscription: "subscription",
  one_time: "one_time",
  refund: "refund",
} as const;

export interface RevenueLog {
  id: number;
  customerId: number;
  customerName: string;
  amount: number;
  type: RevenueLogType;
  occurredAt: string;
}

export interface CustomerDetail {
  customer: Customer;
  lifetimeValue: number;
  recentRevenue: RevenueLog[];
}

export type ListCustomersParams = {
  search?: string;
  status?: ListCustomersStatus;
};

export type ListCustomersStatus =
  (typeof ListCustomersStatus)[keyof typeof ListCustomersStatus];

export const ListCustomersStatus = {
  active: "active",
  inactive: "inactive",
  all: "all",
} as const;

export type ListRevenueLogsParams = {
  limit?: number;
};