# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerStatus } from "./customerStatus";

export interface Customer {
  id: number;
  name: string;
  email: string;
  status: CustomerStatus;
  plan: string;
  mrr: number;
  createdAt: Date;
}