# lib/api-zod/src/generated/types/customerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { RevenueLog } from "./revenueLog";

export interface CustomerDetail {
  customer: Customer;
  lifetimeValue: number;
  recentRevenue: RevenueLog[];
}