# lib/api-zod/src/generated/types/customerStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */

export type CustomerStatus =
  (typeof CustomerStatus)[keyof typeof CustomerStatus];

export const CustomerStatus = {
  active: "active",
  inactive: "inactive",
} as const;