# lib/api-zod/src/generated/types/overviewMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { StatCard } from "./statCard";

export interface OverviewMetrics {
  mrr: StatCard;
  activeUsers: StatCard;
  churnRate: StatCard;
  avgRevenue: StatCard;
}