# lib/api-zod/src/generated/types/planBreakdownItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanBreakdownItem {
  plan: string;
  customers: number;
  revenue: number;
}