# lib/api-zod/src/generated/types/revenueLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { RevenueLogType } from "./revenueLogType";

export interface RevenueLog {
  id: number;
  customerId: number;
  customerName: string;
  amount: number;
  type: RevenueLogType;
  occurredAt: Date;
}