# lib/api-zod/src/generated/types/revenuePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface RevenuePoint {
  /** YYYY-MM */
  month: string;
  /** Short month label e.g. Jan */
  label: string;
  revenue: number;
}