# lib/api-zod/src/generated/types/statCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { StatCardFormat } from "./statCardFormat";
import type { TrendDirection } from "./trendDirection";

export interface StatCard {
  label: string;
  value: number;
  format: StatCardFormat;
  deltaPercent: number;
  trend: TrendDirection;
  positiveIsGood: boolean;
}