# lib/api-zod/src/generated/types/trendDirection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Analytics API
 * OpenAPI spec version: 0.1.0
 */

export type TrendDirection =
  (typeof TrendDirection)[keyof typeof TrendDirection];

export const TrendDirection = {
  up: "up",
  down: "down",
  flat: "flat",
} as const;